Fix Storybook error
