+++ packages/outside/src/PageContainer/Admin/UserListPage/index.tsx
@@ -17,6 +17,7 @@ import { RoleEnumTypes } from '@bitgouel/types'
 import { useRouter } from 'next/navigation'
 import { ChangeEvent, FormEvent, useEffect, useState } from 'react'
 import * as S from './style'
+import { AdminDisplayInfo } from '@/components'
 
 const defaultFilterList = [
   { text: '전체', item: '전체', checked: true },
@@ -103,12 +104,7 @@ const UserListPage = () => {
               <span>필터</span>
             </S.UserSearchFilter>
           </S.UserSearchContainer>
-          <S.DisplayBar>
-            <span>이름</span>
-            <span>직업</span>
-            <span>전화번호</span>
-            <span>이메일</span>
-          </S.DisplayBar>
+          <AdminDisplayInfo />
           <S.UserListContainer>
             {data?.users.map((user) => (
               <UserItem
