+++ packages/outside/src/components/index.ts
@@ -1,2 +1,3 @@
 export { default as ClubItem } from './ClubItem'
 export { default as SchoolFilter } from './SchoolFilter'
+export { default as AdminDisplayInfo } from './AdminDisplayInfo'
