+++ shared/common/src/components/LectureApplyItem/index.stories.tsx
@@ -16,7 +16,7 @@ export const isCompleteItem: Story = {
       id: '1234',
       cohort: 6,
       name: '파티피플공명',
-      school: '광주소프트웨어마이스터고등학교',
+      school: 'GWANGJU_SOFTWARE_MEISTER_HIGH_SCHOOL',
       grade: 3,
       classNumber: 4,
       number: 18,
@@ -34,7 +34,7 @@ export const isNotCompleteItem: Story = {
       id: '1234',
       cohort: 6,
       name: '파티피플공명',
-      school: '광주소프트웨어마이스터고등학교',
+      school: 'GWANGJU_SOFTWARE_MEISTER_HIGH_SCHOOL',
       grade: 3,
       classNumber: 4,
       number: 18,
